Added class-level variables for formatting string responses; Indices of information positions in the provided data array data; Date formatting, whose name had to be shortened to D_FORMATTER due to a build error that limits the string length to 100 characters.
In the getSalaryInfo() method, we first parse the date into LocalDate.
We form part of the response using the above-declared REPORT_TITLE_PATTERN pattern.
And we add the resulting response to the declared StringBuilder.

For further calculations, we add the salary variable.

The first loop iterates over the names array.
Inside the first cycle, the second cycle is triggered, which goes through the elements of the data array.
For each iteration, we split the data element.
If the name in the splitLine matches the name in the name array and the weekday date is in the correct range (this is checked by a separate isBetween() method), then we calculate the salary using the getSalary() method.
We carry out these calculations in the try-catch block in case of a ParseException error.
After exiting the second loop, we create a formattedString string, in which we add data according to the PERSON_DESCRIPTION_PATTERN pattern.
The result is added to StringBuilder.
Reset the salary counter.
At the end of the first loop, remove the extra lineSeparator() and return the response as a String.

The isBetween() method compares whether the date was greater than the start date and less than the end date, or whether it coincides with them.

The getSalary() method parses the values from the income and hours strings and multiplies them with each other.